+++ tests/test_function/test_parse_string.py
@@ -125,6 +125,19 @@ def test_parse_new_guid() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_pad_left() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "pad_left"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path=file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_replace() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "replace"