+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `startsWith` function
 - Support `endsWith` function
 - Support `trim` function
+- Support `padLeft` function
 
 ## 0.2.0 - 2022-02-13